+++ src/isoparticletype.cc
@@ -15,7 +15,7 @@ namespace Smash {
 
 static IsoParticleTypeList iso_type_list;
 
-IsoParticleType::IsoParticleType(const std::string &n, float m, float w,
+IsoParticleType::IsoParticleType(const std::string &n, double m, double w,
                                  unsigned int s)
                                 : name_(n), mass_(m), width_(w), spin_(s) {}
 
@@ -199,12 +199,12 @@ TabulationPtr IsoParticleType::integrate_RR(ParticleTypePtr &type_res_2) {
          type_res_1->min_mass_kinematic() +
          type_res_2->min_mass_kinematic(),
          3.f, 125,
-         [&](float srts) {
+         [&](double srts) {
             const auto result = integrate2d(type_res_1->min_mass_kinematic(),
                                srts - type_res_2->min_mass_kinematic(),
                                type_res_2->min_mass_kinematic(),
                                srts - type_res_1->min_mass_kinematic(),
-                               [&](float m1, float m2) {
+                               [&](double m1, double m2) {
                                   return spec_func_integrand_2res(srts, m1, m2,
                                                       *type_res_1, *type_res_2);
                                });
